move timer elapsed logic to timer strategy
