+++ Raft.Test/Strategy/Timer/LeaderTimerStrategyShould.cs
@@ -12,19 +12,22 @@ namespace Raft.Test.Strategy.Timer
     {
         private Node _leader;
         private LeaderTimerStrategy _leaderTimerStrategy;
+        private IMessageBroker _messageBroker;
 
         [SetUp]
         public void SetUp()
         {
-            _leader = new Node("Leader", Substitute.For<IMessageBroker>());
+            _messageBroker = Substitute.For<IMessageBroker>();
+            _leader = new Node("Leader", _messageBroker)
+            {
+                Status = new LeaderStatus(0)
+            };
             _leaderTimerStrategy = new LeaderTimerStrategy(_leader);
         }
         
         [Test]
         public void ResetTimer_WhenMessageFromNewerLeaderReceived()
         {
-            _leader.Status = new LeaderStatus(0);
-
             var nodeMessage = new NodeMessage(1, "ping", MessageType.Info, "anotherLeader", Guid.Empty);
             var shouldReset = _leaderTimerStrategy.ShouldReset(nodeMessage);
             
@@ -40,6 +43,14 @@ namespace Raft.Test.Strategy.Timer
             var shouldReset = _leaderTimerStrategy.ShouldReset(nodeMessage);
             
             shouldReset.ShouldBeFalse();
+        }
+        
+        [Test]
+        public void Ping_OnTimerElapsed()
+        {
+            _leaderTimerStrategy.OnTimerElapsed();
+            
+            _messageBroker.Received(1).Broadcast(Arg.Is<NodeMessage>(m => m.Type == MessageType.Info));
         }
     }
 }
