+++ Raft/NodeStrategy/FollowerStrategy.cs
@@ -50,11 +50,6 @@ namespace Raft.NodeStrategy
             }
         }
 
-        public void OnTimerElapsed()
-        {
-            Node.BecomeCandidate();
-        }
-
         private void Vote(NodeMessage message)
         {
             Node.Vote(message.Term, message.SenderName, message.Id);
